refactor: move CredentialFilter back to credential/view.py
Relates to JIRA: DISCOVERY-797
